[Fix] 오래된 로그인 유저 제한 기준 일자 오버플로우 버그 수정

+++ app/src/main/java/com/example/kwoncheolhyeok/core/Util/DataContainer.java
@@ -13,7 +13,7 @@ public class DataContainer {
     public static final int ChildrenMax = 1000;
     public static final int ViewedMeMax = 45;
     public static final int CoreCloudMax = 7;
-    public static final int SecToDay = TIME_MAXIMUM.SEC*TIME_MAXIMUM.MIN*TIME_MAXIMUM.HOUR*1000;
+    public static final long SecToDay = TIME_MAXIMUM.SEC*TIME_MAXIMUM.MIN*TIME_MAXIMUM.HOUR*1000;
 //    public static final int SecToDay = TIME_MAXIMUM.SEC*TIME_MAXIMUM.MIN*TIME_MAXIMUM.HOUR*1000/24/60;
 
     private static class TIME_MAXIMUM {
